Refactor merkle distribution script dependencies
