+++ src/scripts/pre-rewards/rewards.js
