+++ src/scripts/tbtcv2-rewards/tsconfig.json
